packages/keyword-optimiser/requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiomultiprocess==0.9.0
aiosignal==1.2.0
argcomplete==2.0.0
async-timeout==4.0.2
attrs==21.4.0
bc-python-hcl2==0.3.42
beautifulsoup4==4.11.1
boto3==1.24.13
botocore==1.27.13
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
checkov==2.0.1226
click==8.1.3
click-option-group==0.5.3
cloudsplaining==0.5.0
colorama==0.4.5
ConfigArgParse==1.5.3
contextlib2==21.6.0
cyclonedx-python-lib==2.6.0
decorator==5.1.1
deep-merge==0.0.4
detect-secrets==1.2.0
docker==5.0.3
dockerfile-parse==1.2.0
dpath==1.5.0
frozenlist==1.3.0
fuzzywuzzy==0.18.0
gitdb==4.0.9
GitPython==3.1.27
google-ads==17.0.0
google-api-core==2.8.2
google-auth==1.35.0
google-auth-oauthlib==0.5.2
googleads==32.0.0
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.5.3
jsonschema==3.2.0
junit-xml==1.9
lark==1.1.2
lxml==4.9.0
Markdown==3.3.7
MarkupSafe==2.1.1
multidict==6.0.2
mysql-connector-python==8.0.29
networkx==2.8.4
numpy==1.23.0
oauthlib==3.2.0
packageurl-python==0.9.9
packaging==21.3
pandas==1.4.3
platformdirs==2.5.2
ply==3.11
policy-sentry==0.12.3
policyuniverse==1.5.0.20220523
prettytable==3.3.0
proto-plus==1.19.6
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycep-parser==0.3.7
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytorch-transformers==1.1.0
pytz==2022.1
PyYAML==5.4.1
regex==2022.6.2
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
s3transfer==0.6.0
schema==0.7.5
scipy==1.8.1
semantic-text-similarity==1.0.3
semantic-version==2.10.0
sentencepiece==0.1.96
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
strsim==0.0.3
tabulate==0.8.10
termcolor==1.1.0
toml==0.10.2
torch==1.12.0
tqdm==4.64.0
typing_extensions==4.2.0
update-checker==0.18.0
urllib3==1.26.9
wcwidth==0.2.5
websocket-client==1.3.3
xmltodict==0.13.0
yarl==1.7.2
zeep==4.1.0
zipp==3.8.0
